requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
python-Levenshtein
aio_pika
python-dotenv
